addition of static method and login fuction refactored
