+++ src/routes/book.js
@@ -1,7 +1,7 @@
 import express from 'express';
 const Router = express.Router();
 import { getAllBooks, getOneBook, createBook, removeBook } from '../controller/book.js';
-import validateCreateBook from '../utils/bookValidation.js';
+import validateCreateBook from '../middleware/validator/bookValidation.js';
 import { librarianAuth } from '../middleware/auth.js';
 
 Router.get('/', getAllBooks);
